Add support for Picture upload with DnD support.
Add preliminary support for uploading a picture for the artists.
There are still a lot of rough edges but I felt it was worth
making it available to see.

* Add blueimp jQuery File uploader.
* Add picture file upload element to artist form.
* Display the Artist's picture in the view page.
* Add Picture model.
* Add Filestore library for storing files in a zip
    file. Without locking this is bound to !@# up.
    I will be reverting this change.

The decision to add a Zip file store was made to
attempt to avoid file inclusion attacks. To properly
pursue this idea locking would be required.

In a future version I will add an IoC container and
try it again.